[WIP] trying to componentize Explore
